+++ src/main/java/com/hexvane/strangematter/block/EnergeticShardOreBlock.java
@@ -1,7 +1,5 @@
 package com.hexvane.strangematter.block;
 
-import com.hexvane.strangematter.StrangeMatterMod;
-
 /**
  * Energetic Shard Ore Block - drops energetic shards when mined.
  * Spawns beneath energetic rift anomalies.
@@ -9,6 +7,6 @@ import com.hexvane.strangematter.StrangeMatterMod;
 public class EnergeticShardOreBlock extends ShardOreBlock {
     
     public EnergeticShardOreBlock() {
-        super(StrangeMatterMod.ENERGETIC_SHARD);
+        super();
     }
 }