+++ src/main/java/com/hexvane/strangematter/block/GraviticShardOreBlock.java
@@ -1,8 +1,5 @@
 package com.hexvane.strangematter.block;
 
-import com.hexvane.strangematter.StrangeMatterMod;
-import net.minecraftforge.registries.RegistryObject;
-
 /**
  * Gravitic Shard Ore Block - drops gravitic shards when mined.
  * Spawns beneath gravity anomalies.
@@ -10,6 +7,6 @@ import net.minecraftforge.registries.RegistryObject;
 public class GraviticShardOreBlock extends ShardOreBlock {
     
     public GraviticShardOreBlock() {
-        super(StrangeMatterMod.GRAVITIC_SHARD);
+        super();
     }
 }